requirements.txt
streamlit
google-generativeai
pytest
